Fix some integer types

+++ instructions.c
@@ -252,7 +252,7 @@ void run_ld_b_vx(struct chip8 *vm, uint16_t instruction) {
 void run_ld_i_vx(struct chip8 *vm, uint16_t instruction) {
     uint8_t reg = REG_1(instruction);
     uint16_t location = vm->reg_i;
-    int i;
+    uint8_t i;
     for (i = 0; i <= reg; ++i) {
         vm->ram[location++] = vm->reg_v[i];
     }
@@ -261,7 +261,7 @@ void run_ld_i_vx(struct chip8 *vm, uint16_t instruction) {
 void run_ld_vx_i(struct chip8 *vm, uint16_t instruction) {
     uint8_t reg = REG_1(instruction);
     uint16_t location = vm->reg_i;
-    int i;
+    uint8_t i;
     for (i = 0; i <= reg; ++i) {
         vm->reg_v[i] = vm->ram[location++];
     }
